Fix not to jump to the declartion but the function itself.
Commit 45d2980b does not fix it fully.

It is possible that the function declaration does not contain a semicolon in
the end of the first line (for example, arguments are written in seperated
lines).
In such cases, text-to-search-for matches for both the declaration and the
function itself, search will stop at the declartion.

This commit works around for these cases. It searches backward instead. Thus
it only works if the selected tag is the only tag in the file. The better
solution would be to refer the line number in the TAGS file. (TODO)